webapp/requirements.txt
gunicorn
schedule==0.6.0
django-heroku==0.3.1
dj-database-url==0.5.0
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.4.4
beautifulsoup4==4.9.1
captcha==0.3
certifi==2019.3.9
chardet==3.0.4
comtypes==1.1.7
defusedxml==0.6.0
distlib==0.3.0
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.3.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-user-agents==0.4.0
docopt==0.6.2
filelock==3.0.12
idna==2.8
joblib==0.14.1
Js2Py==0.70
numpy==1.18.4
oauthlib==3.0.1
packaging==20.4
pandas==1.0.3
pep8==1.7.1
pillow==7.1.2
pipwin==0.5.0
pycodestyle==2.5.0
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.2
pySmartDL==1.3.3
python-dateutil==2.8.1
python-memcached==1.59
python3-openid==3.1.0
pyttsx3==2.87
PyYAML==5.3.1
requests==2.21.0
requests-oauthlib==1.2.0
shortuuid==1.0.1
six==1.14.0
soupsieve==2.0.1
sqlparse==0.2.4
threadpoolctl==2.0.0
tzlocal==2.1
ua-parser==0.10.0
urllib3==1.24.2
user-agents==2.1
virtualenv==20.0.21
psycopg2==2.8.5
django-redirect-to-non-www==0.1.1
django-compression-middleware==0.4.1